// src/app/md-module/services/mrc/model/pageMedicalRecordRequest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MedicalRecordRequest } from './medicalRecordRequest';
import { Pageable } from './pageable';
import { Sort } from './sort';


export interface PageMedicalRecordRequest { 
    content?: Array<MedicalRecordRequest>;
    empty?: boolean;
    first?: boolean;
    last?: boolean;
    number?: number;
    numberOfElements?: number;
    pageable?: Pageable;
    size?: number;
    sort?: Sort;
    totalElements?: number;
    totalPages?: number;
}